@tzneal 2024-01-31
Let's run a benchmark of the scheduler before/after this change and compare just to check performance.