@jmdeal 2024-06-13
Benchmarking results:
```
goos: darwin
goarch: arm64
pkg: sigs.k8s.io/karpenter/pkg/controllers/provisioning/scheduling
                  │ old_results.txt │           new_results.txt            │
                  │     sec/op      │    sec/op      vs base               │
Scheduling1-10         99.67µ ± 42%   103.39µ ± 30%       ~ (p=0.989 n=20)
Scheduling50-10        305.6µ ±  6%    305.5µ ±  7%       ~ (p=0.265 n=20)
Scheduling100-10       488.9µ ±  6%    451.0µ ±  5%  -7.76% (p=0.001 n=20)
Scheduling500-10       1.929m ±  1%    1.859m ±  2%  -3.66% (p=0.002 n=20)
Scheduling1000-10      3.955m ±  3%    3.999m ±  7%       ~ (p=0.478 n=20)
Scheduling2000-10      7.518m ±  7%    8.072m ±  4%  +7.38% (p=0.013 n=20)
Scheduling5000-10      19.45m ±  9%    17.96m ±  4%  -7.69% (p=0.038 n=20)
geomean                1.494m          1.477m        -1.13%

                  │ old_results.txt │           new_results.txt           │
                  │      nodes      │   nodes     vs base                 │
Scheduling1-10           1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=20) ¹
Scheduling50-10          5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=20) ¹
Scheduling100-10         5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=20) ¹
Scheduling500-10         5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=20) ¹
Scheduling1000-10        5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=20) ¹
Scheduling2000-10        5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=20) ¹
Scheduling5000-10        5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=20) ¹
geomean                  3.973        3.973       +0.00%
¹ all samples are equal

                  │ old_results.txt │           new_results.txt            │
                  │      pods       │    pods      vs base                 │
Scheduling1-10          1.000 ±  0%   1.000 ±  0%       ~ (p=1.000 n=20) ¹
Scheduling50-10         29.00 ± 17%   29.00 ± 24%       ~ (p=0.814 n=20)
Scheduling100-10        25.50 ± 25%   26.00 ± 23%       ~ (p=0.732 n=20)
Scheduling500-10        28.00 ± 50%   25.00 ± 32%       ~ (p=0.587 n=20)
Scheduling1000-10       35.50 ± 21%   29.00 ± 21%       ~ (p=0.279 n=20)
Scheduling2000-10       29.50 ± 39%   28.00 ± 29%       ~ (p=0.815 n=20)
Scheduling5000-10       34.00 ± 41%   31.00 ± 35%       ~ (p=0.344 n=20)
geomean                 18.48         17.36        -6.10%
¹ all samples are equal

                  │ old_results.txt │           new_results.txt           │
                  │    pods/sec     │   pods/sec    vs base               │
Scheduling1-10        10.044k ± 30%   9.672k ± 24%       ~ (p=0.984 n=20)
Scheduling50-10        163.6k ±  6%   163.7k ±  8%       ~ (p=0.265 n=20)
Scheduling100-10       204.5k ±  6%   221.8k ±  5%  +8.42% (p=0.001 n=20)
Scheduling500-10       259.2k ±  1%   269.0k ±  2%  +3.80% (p=0.002 n=20)
Scheduling1000-10      252.9k ±  3%   250.0k ±  8%       ~ (p=0.478 n=20)
Scheduling2000-10      266.1k ±  7%   247.8k ±  4%  -6.91% (p=0.013 n=20)
Scheduling5000-10      257.0k ± 10%   278.4k ±  3%  +8.32% (p=0.038 n=20)
geomean                147.3k         149.0k        +1.12%
```